fix: Allow empty generated pipelines to be skipped
If a generated pipeline has no steps generated, whether from the watched paths or via aspects of the plugin ie hooks, it skips attempting to upload it. This is similar to the behaviour if there is no diff output.

In order to accomplish this, added a boolean to the output of generatePipeline() returning true if the generated pipeline has at least 1 step defined else returns false. This is then checked in uploadPipeline() and if the returned value was false ie there were no generated steps, uploadPipeline() returns without attempting to upload the empty pipeline.